newspulselast/newspulse/requirements.txt
streamlit>=1.36.0
pymongo[srv]>=4.7.0
python-dotenv>=1.0.1
spacy>=3.7.4
nltk>=3.8.1
transformers>=4.41.2
torch>=2.3.0
datasets>=2.20.0
pandas>=2.2.2
matplotlib>=3.8.4
seaborn>=0.13.2
plotly>=5.22.0
bcrypt>=4.1.3
tqdm>=4.66.4
orjson>=3.10.6
feedparser>=6.0.11
requests>=2.32.3

google-generativeai>=0.7.2

